requirements.txt
requests
beautifulsoup4
langchain
langchain-openai
python-dotenv
tweepy
openai
faiss-cpu
pytz
langgraph>=0.2.0
pydantic[email]
textblob==0.18.0
nltk==3.8.1
streamlit
matplotlib
flask
werkzeug
vaderSentiment
cloudinary